# docs/swagger/swagger.yaml
basePath: /gateway
definitions:
...
  github_com_ftryyln_hotel-booking-microservices_pkg_dto.BookingRequest:
    properties:
      check_in:
        $ref: '#/definitions/github_com_ftryyln_hotel-booking-microservices_pkg_dto.Date'
      check_out:
        $ref: '#/definitions/github_com_ftryyln_hotel-booking-microservices_pkg_dto.Date'
      guests:
        type: integer
      room_type_id:
        type: string
      user_id:
        type: string
    type: object
  github_com_ftryyln_hotel-booking-microservices_pkg_dto.BookingResponse:
    properties:
      check_in:
        type: string
      check_out:
        type: string
      guests:
        type: integer
      id:
        type: string
      payment:
        $ref: '#/definitions/github_com_ftryyln_hotel-booking-microservices_pkg_dto.PaymentResponse'
      status:
        type: string
      total_nights:
        type: integer
      total_price:
        type: number
    type: object
  github_com_ftryyln_hotel-booking-microservices_pkg_dto.CheckpointRequest:
    properties:
      action:
        type: string
    type: object
...
  github_com_ftryyln_hotel-booking-microservices_pkg_dto.Date:
    properties:
      time.Time:
        type: string
    type: object
  github_com_ftryyln_hotel-booking-microservices_pkg_dto.ErrorResponse:
    properties:
      code:
        type: string
      details: {}
      message:
        type: string
    type: object
...
  github_com_ftryyln_hotel-booking-microservices_pkg_dto.PaymentResponse:
    properties:
      id:
        type: string
      payment_url:
        type: string
      provider:
        type: string
      status:
        type: string
    type: object
...
host: localhost:8088
info:
  contact: {}
  description: Aggregated API surface for hotel booking platform.
  title: Hotel Booking Microservices API
  version: "1.0"
paths:
...
  /bookings:
    get:
      parameters:
      - description: pagination limit (default 50)
        in: query
        name: limit
        type: integer
      - description: pagination offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_ftryyln_hotel-booking-microservices_pkg_dto.BookingResponse'
            type: array
      security:
      - BearerAuth: []
      summary: List bookings
      tags:
      - Bookings
    post:
      consumes:
      - application/json
      parameters:
      - description: Booking payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_ftryyln_hotel-booking-microservices_pkg_dto.BookingRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_ftryyln_hotel-booking-microservices_pkg_dto.BookingResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_ftryyln_hotel-booking-microservices_pkg_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create booking
      tags:
      - Bookings
  /bookings/{id}:
    get:
      parameters:
      - description: Booking ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_ftryyln_hotel-booking-microservices_pkg_dto.BookingResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_ftryyln_hotel-booking-microservices_pkg_dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_ftryyln_hotel-booking-microservices_pkg_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get booking
      tags:
      - Bookings
  /bookings/{id}/cancel:
    post:
      parameters:
      - description: Booking ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_ftryyln_hotel-booking-microservices_pkg_dto.BookingResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_ftryyln_hotel-booking-microservices_pkg_dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_ftryyln_hotel-booking-microservices_pkg_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Cancel booking
      tags:
      - Bookings
  /bookings/{id}/checkpoint:
    post:
      consumes:
      - application/json
      parameters:
      - description: Booking ID
        in: path
        name: id
        required: true
        type: string
      - description: Checkpoint payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_ftryyln_hotel-booking-microservices_pkg_dto.CheckpointRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_ftryyln_hotel-booking-microservices_pkg_dto.BookingResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_ftryyln_hotel-booking-microservices_pkg_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Booking checkpoint
      tags:
      - Bookings
  /bookings/{id}/status:
    get:
      parameters:
      - description: Booking ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_ftryyln_hotel-booking-microservices_pkg_dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_ftryyln_hotel-booking-microservices_pkg_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get booking status
      tags:
      - Bookings
    post:
      consumes:
      - application/json
      parameters:
      - description: Booking ID
        in: path
        name: id
        required: true
        type: string
      - description: Status payload
        in: body
        name: request
        required: true
        schema:
          additionalProperties:
            type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_ftryyln_hotel-booking-microservices_pkg_dto.BookingResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_ftryyln_hotel-booking-microservices_pkg_dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_ftryyln_hotel-booking-microservices_pkg_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Change booking status (admin)
      tags:
      - Bookings
...
schemes:
- http
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"